Added Notes database. Note POJO was moved.
