--- app/src/main/java/com/dvinc/notepad/data/Note.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2018 by Denis Verentsov (decsent@Yandex.ru)
- * All rights reserved.
- */
-
-package com.dvinc.notepad.data
-
-data class Note(
-        val name: String,
-        val content: String,
-        val updateTime: Long)
