+++ app/src/main/java/com/dvinc/notepad/data/database/dao/Note.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2018 by Denis Verentsov (decsent@Yandex.ru)
+ * All rights reserved.
+ */
+
+package com.dvinc.notepad.data.database.dao
+
+import android.arch.persistence.room.ColumnInfo
+import android.arch.persistence.room.Entity
+import android.arch.persistence.room.PrimaryKey
+
+@Entity(tableName = "Notes")
+data class Note(
+        @PrimaryKey(autoGenerate = true) var id: Long = 0,
+        @ColumnInfo(name = "name") val name: String,
+        @ColumnInfo(name = "content") val content: String,
+        @ColumnInfo(name = "update_time") val updateTime: Long)
